Mostly completed separation of editor
Need to separate views and models for REPL, resolve conflicts with master, and then complete separation
